Refactor dataflows
