+++ tests/tgan/test_gan.py
@@ -1,11 +1,11 @@
-from unittest import TestCase, expectedFailure
+from unittest import expectedFailure
 from unittest.mock import MagicMock, patch
 
 import tensorflow as tf
 from tensorflow.test import TestCase as TensorFlowTestCase
 from tensorpack.tfutils.tower import TowerFuncWrapper
 
-from tgan.gan import GANModelDesc, GANTrainer, MultiGPUGANTrainer, RandomZData, SeparateGANTrainer
+from tgan.gan import GANModelDesc, GANTrainer, MultiGPUGANTrainer, SeparateGANTrainer
 
 
 class TestGanModelDesc(TensorFlowTestCase):
@@ -287,35 +287,3 @@ class TestMultiGpuGanTrainer(TensorFlowTestCase):
             # Check
             assert len(error.args) == 1
             assert error.args[0] == 'nr_gpu must be strictly greater than 1.'
-
-
-class TestRandomZData(TestCase):
-
-    @patch('tgan.gan.DataFlow.__init__', autospec=True)
-    def test___init__(self, dataflow_mock):
-        """On init, shape is set as attribute and super is called."""
-        # Setup
-        shape = (10, 2)
-
-        # Run
-        instance = RandomZData(shape)
-
-        # Check
-        assert instance.shape == (10, 2)
-        dataflow_mock.assert_called_once_with()
-
-    @patch('tgan.gan.np.random.normal', autospec=True)
-    def test_get_data(self, normal_mock):
-        """get_data return an infinite generator of normal vectors of the given shape."""
-        # Setup
-        shape = (2, 1)
-        instance = RandomZData(shape)
-        normal_mock.return_value = [[0.5], [0.5]]
-
-        # Run
-        generator = instance.get_data()
-        result = next(generator)
-
-        # Check
-        assert result == [normal_mock.return_value]
-        normal_mock.assert_called_once_with(0, 1, size=(2, 1))
